removed some fixes and refactored code using services

+++ app/controllers/subscriptions_controller.rb
@@ -3,8 +3,8 @@ class SubscriptionsController < ApplicationController
   before_action :make_payment, only: %i[create]
   after_action :perform_transaction, only: %i[success]
   def create
+    authorize Subscription
     @subscription = Subscription.new(subscription_params)
-    authorize @subscription
   end
   def success
     @subscription = Subscription.new(subscription_params)
@@ -16,6 +16,7 @@ class SubscriptionsController < ApplicationController
   end
 
   def destroy
+    authorize Subscription
     if @subscription.destroy
       redirect_to plans_url, notice: 'Unsubscribed.'
     else
@@ -37,22 +38,24 @@ class SubscriptionsController < ApplicationController
   end
 
   def perform_transaction
-    transaction = Transaction.create(user_id: @subscription.user_id, subscription_id: @subscription.id, amount: @subscription.plan.monthly_fee)
-    ReceiptMailer.with(user: @subscription.user, transaction: transaction).transaction_created.deliver_later
+    transaction = Transaction.new(@subscription)
+    if transaction.perform_transaction
+      ReceiptMailer.with(user: @subscription.user, transaction: transaction).transaction_created.deliver_later
+    end
   end
 
   def make_payment
     plan = Plan.find(params[:plan_id])
-    @session = Stripe::Checkout::Session.create({
-      customer: current_user.stripe_customer_id,
-      payment_method_types: ['card'],
-      line_items: [
-        price: plan.stripe_price_id,
-        quantity: 1,
-      ],
-      mode: 'subscription',
-      success_url: subscription_success_url(user_id: params[:user_id], plan_id: params[:plan_id]),
-      cancel_url: root_url,
-    })
+    unless current_user.stripe_customer_id
+      success_url = subscription_success_url(user_id: params[:user_id], plan_id: params[:plan_id])
+      session = CheckoutSession.new(plan, current_user, success_url)
+      @session = session.create
+    else
+      flash[:alert] = "Unable to subscribe"
+      redirect_to root_url
+    end
+  rescue ActiveRecord::RecordNotFound
+    flash[:alert] = "This plan doesn't exist!"
+    redirect_to root_url
   end
 end
